+++ src/rai_core/rai/communication/ros2/connectors/__init__.py
@@ -12,10 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .action_mixin import ROS2ActionMixin
 from .ari_connector import ROS2ARIConnector
 from .hri_connector import ROS2HRIConnector
+from .service_mixin import ROS2ServiceMixin
 
 __all__ = [
     "ROS2ARIConnector",
+    "ROS2ActionMixin",
     "ROS2HRIConnector",
+    "ROS2ServiceMixin",
 ]
